Modified for-loops that loop on stl containers to use 'size_t' instead of 'int' for the loop variable, in order to remove 'signed/unsigned integer comparison' compiler warnings. Modified xml code generation tools along the same lines and regenerated all state Attributes and Python attributes which did not contain un-regenerateable specialized code.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@4154 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/Database/Database/avtDatasetCollection.C
@@ -265,7 +265,7 @@ avtDatasetCollection::AddMixVar(int dom, void_ref_ptr mv)
 void_ref_ptr
 avtDatasetCollection::GetMixVar(int dom, std::string name)
 {
-    for (int i = 0 ; i < mixvars[dom].size() ; i++)
+    for (size_t i = 0 ; i < mixvars[dom].size() ; i++)
     {
         avtMixedVariable *mixvar = (avtMixedVariable *) *mixvars[dom][i];
         if (mixvar->GetVarname() == name)
@@ -321,7 +321,7 @@ avtDatasetCollection::ReplaceMixVar(int dom, void_ref_ptr mix)
 {
     avtMixedVariable *mv = (avtMixedVariable *) *mix;
 
-    for (int i = 0 ; i < mixvars[dom].size() ; i++)
+    for (size_t i = 0 ; i < mixvars[dom].size() ; i++)
     {
         avtMixedVariable *nmv = (avtMixedVariable *) *mixvars[dom][i];
         if (nmv->GetVarname() == mv->GetVarname())
